+++ spec/storey/excluded_models_spec.rb
@@ -2,7 +2,10 @@ require 'spec_helper'
 
 describe Storey, "dealing with excluded_models" do
   before do
-    Storey.excluded_models = %w(Company)
+    Storey.configuration.excluded_models = %w(Company)
+    # In practice, `excluded_models` is not set on the fly, but since we do so
+    # in the tests, we must call `init` to set the models to the right tables:
+    Storey.init
   end
 
   it "should always reference these models in the public schema" do